separated general and music functions into their own classes
