+++ spoti.py
@@ -1,5 +1,5 @@
-import requests
 import base64
+import requests
 
 
 class SpotifyAPI:
